Refactor: Move validation into Attribute:: namespace
