+++ lib/model_configuration/attribute/validation.rb
@@ -1,0 +1,19 @@
+class ModelConfiguration::Attribute::Validation
+
+  attr_reader :attribute, :key, :args
+
+  # EG:
+  # attribute: name
+  # key: presence
+  # args: true OR {message: "YOLO"}
+  def initialize(attribute, key, args)
+    @attribute = attribute
+    @key = key
+    @args = args
+  end
+
+  def implementation
+    "#{key}: #{args}"
+  end
+
+end
