+++ lib/model_configuration/model_configuration.rb
@@ -1,7 +1,6 @@
 require_relative "attribute/factory.rb"
 require_relative "attribute.rb"
 require_relative "url_builder.rb"
-require_relative "validation.rb"
 
 class ModelConfiguration
